convert neth-lib.js to .ts
