+++ packages/neth/tsconfig.lib.json
@@ -5,6 +5,6 @@
     "declaration": true,
     "types": []
   },
-  "include": ["**/*.ts", "src/lib/neth-lib.js"],
+  "include": ["**/*.ts", "src/lib/neth-lib.ts"],
   "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
 }
